requirements.txt
open-jarvis-sdk
websocket-client

# apt install python3-pyaudio -y